requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.21.1
python-dotenv==1.0.0
cloudinary==1.31.0
requests==2.31.0
numpy==1.26.0
opencv-python==4.8.0.76
ultralytics==8.0.134
pydantic==1.10.9
